fix: Pass the correct extensions
It looks like `extensions` was created with the intention to pass the extensions to the Column constructor. However, a second variable `extentions` was also created to store the extensions identified in the definition args. Unfortunately this second variable was not used, despite containing the desired extensions, and the original variable, which contained the unfiltered definition, was passed to the constructor.

Instead, the variable containing the unfiltered args has been removed, and the typo in the variable containing the filtered list has been corrected. This results in the correct extensions being passed to the Column constructor.