fix: Support time zone when serializing DateTimes
Previously a DateTime being serialized would ignore the timezone, effectively assuming the timezone of the value was the same as the column.

Instead, the timezone is now taken into account so the time can be given in any timezone and it will be converted to the timezone of the column.