fix: Make BooleanType handle 0 correctly
The tests BooleanType had a typo causing them to skip testing serialization of and casting of 0. This meant that even though it treated 0 as a truthy value, it was not caught by the tests.

Instead, BooleanType now uses a list of truthy values for serialization, which now includes 0. This fixes the handling, and the tests have been updated to enforce this behaviour.